starts working on this, mostly done
