+++ lib/src/extensions/when_orientation.dart
@@ -1,0 +1,15 @@
+part of 'extensions.dart';
+
+Widget _$whenOrientationLandscape(
+        Widget landscapeWidget, Widget originalWidget) =>
+    ResponsiveByOrientation(
+      landscapeChild: landscapeWidget,
+      portraitChild: originalWidget,
+    );
+
+Widget _$whenOrientationPortrait(
+        Widget portraitWidget, Widget originalWidget) =>
+    ResponsiveByOrientation(
+      landscapeChild: originalWidget,
+      portraitChild: portraitWidget,
+    );
